linter,meta: fix panic during isset($$x)
Add support for variable-variable tracking inside a scope
as well, so we don't mark usages of $$x inside if body
as undefined. isset($$x) defines $x along $$x, so usages
of $x are permitted inside that body as well.
Doesn't work for arbitrary indirection level,
isset($$$x) is not supported (but it won't panic, as it did before).

Updates #26

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>