cmd: add -allow-checks argument
Makes possible to run some checks in a separation,
which is useful in CI and while testing a new check.

Also makes it possible to have "disabled by default" checks
that are useful, but not for every linter user. It's also
easier to leave some checks disabled by default if they have
some correctness issues that can't be solved in a short period of time.

To figure out what checks are enabled by default we introduce
structured diagnostics info, linter.CheckInfo.
It's not mandatory to register a check with linter.DeclareCheck,
but it's the only way to associate metadata with diagnostics
implemented by any checker.

Non-declared checks can't be enabled by default, so
explicit -allow-checks entry should be specified for them.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>