linter: add callStatic check
Detect when static methods are called using "->" syntax
and instance methods that are called using "::" syntax.

It's not always a mistake, but could be a sign of a code smell.
Also, it makes code harder to read, since it's not apparent
what happens before called method signature is reviewed.

Added "Static" field to meta.FuncInfo to make this check possible.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>