linter: add caseContinue check
Detects "continue" (without level) usage inside switch
that behaves identically to "break".

As a consequence, it's either a bad style (if it's used
instead of break on purpose) or a bug (if "continue"
was intended, if there is a "for" loop outside).
Since both cases are bad, we report any continue (w/o level)
usage inside switch cases.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>