linter: add Filename() to the BlockContext
Sometimes block-scoped checks need to know the filename as well.

Removed "full name" promise from the documentation as it may not
be the case if linter input was not created inside "cmd", which
does map paths to abs paths. In tests we don't do this, so
Filename() does not return the full name inside them.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>