all: use inputs.SourceInput for PHP reading
Add inputs package that defines SourceInput interface.
That interface is used inside linter to read PHP files.

Readers returned from that interface should yield UTF-8 encoded
data. If other encodings are needed, it's possible to handle
them inside linter extensions by providing custom implementation
and assigning it to linter.SrcInput.

Example that adds win-1251 encoding support will be added soon
to show how it can be done.

This is experimental part of the linter and can be a subject of
changes in future. Only UTF-8 support is guaranteed.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>