accept unused-var-regex option
Makes it possible to set "_" as a valid prefix for unused variables.

Useful for tuple-like returns that are destructured using list.

Only one value is used:
	list($_id, $_name, $email) = getUser(...);

If $_ is used everywhere, it's less opparent what fields are ignored:
	list($_, $_, $email) = getUser(...);

The next step is to warn on usages of variables that are matched
by IsDiscardVar. If such variable is read, it could be a mistake.

By default, only "$_" is permitted, as before.
For arbitrary patterns regexp is used.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>