inherit base method param types
If no other type info about params is available,
walk base types to find matching method and see whether
it has proper types info for that parameter.

Do this only for interfaces.

This makes it possible to resolve both input and output
types of methods that implement some interface. In turn,
this makes "undefined method" go away if method had no
PHPDoc that specified the parameter type that was used.

Phpstorm and phpdoc utility do more or less the same thing.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>