Handle late static binding
This change enables more precise handling of `$this` and `static`
types that are quite common in Fluent APIs.

The idea is to carry the current class type along the type resolving
and map "static" pseudo-type into concrete class type.

If current class type is unknown, we can't resolve static to anything
better than "self".

Fixes #89

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>