fix false positives with instanceof
When instanceof is inside if condition, it introduces
a new tracked var via b.addVar().
It seems to be incorrect for cases where added variable
was already used above.
Given that we enter (walk) instanceof LHS before
we add that new variable, instanceof itself doesn't
end up marking that variable as used.

This change makes linter stop tracking instanceof LHS.
(We add it to the scope directly instead.)

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>