add custom meta cacher support
If a custom root checker records some info during the
indexing phase that is required during the analysis, it
must be written to cache, because otherwise that
checker is not going to be executed if cache exists while
that cache would still not contain relevant data.

Make it possible to register a checker along with
associated MetaCacher that can write needed parts
to the cache files and load that info during
cache file reading.

MetaCacher also includes a Version() method
that can be used to implement cache invalidation.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>