change combined cache file layout
If there are no custom checks, cache file looks like this:

	[ver][data]

Where data is gob-encoded.

Previous format for combined cache looked like this:

	[ver][data][x-ver-len1][x-ver1][x-data1]...[x-ver-lenN][x-verN][x-dataN]

Where each data is expected to be gob-encoded (but it's not a restriction).

This resulted in data corruptions due to the mix of write-byte and gob-encode
operations through shared reader/writer.

Potentially better format packs all versions in the header:

	[ver][x-ver-len1][x-ver1]...[x-ver-lenN][x-verN][data][x-data1]...[x-dataN]
        |                                             /
        header---------------------------------------/

It also makes version check faster, we don't need to call custom Decode
on the outdated cache file data.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>