improve type inference
An attempt to improve type inference.
Currently, if an expression has type T, it doesn't mean:
	"expression always have type T"
it only states that
	"expressions might have type T in some cases"

So, the function below is believed to return "int" type:

	function f($cond) {
		if ($cond) { return 10; }
		return $cond;
	}

The bad part is that it can return a value of type($cond).
The right answer would be `int|type($cond)`, but `int|mixed`
would still be better for now.

- try to fix misspelled types instead of just complaining about them

- make functions/expressions always have type.
  When precise type can't be inferred, use void/mixed.
  To avoid tons of new allocations, void and mixed type maps are shared
  as immutable type maps.

- move returnTypes from block context into the block itself,
  since now block checker lives for the entire function scope.
  It makes messing up with merging context impossible.
  We had several issues with type inference due to that
  (fixed in this change, tests are included).

- add left/right shifts type inference

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>