all: add dynamic rules support (initial implementation)
Added a new command-line argument, -rules=<files-list>.
When this argument is set, NoVerify parses these files
and makes dynamic rules out of it. Dynamic rules are a way
to add diagnostics to NoVerify without a need to write any Go
code or even re-compile the linter.

The rules syntax and matching engine is borrowed from
	https://github.com/quasilyte/phpgrep

No measurable performance impact when rules are not used.
When rules are used, we use rule clusterization to
reduce the amount of excessive work performed.
Some more optimizations can be applied later if we find
any bottlenecks or if there is a demand to make this mechanism faster.
Right now it can execute hundreds of rules without significant
penalties (unless all rules belong to the same category,
which can be the case with function call patterns, but that
specific case can be easily optimized in separation).

TODO: add more documentation, tests, and examples.

Right now this feature is not mentioned in docs since
it needs to be better tested prior to that.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>