rules: make it possible to have several patterns for 1 phpdoc comment
When several patterns have exactly the same phpdoc comment, we want
to refactor that or to re-use constraints in some way.

Proposed way for that is phpdoc combining.

If a pattern doesn't have its own phpdoc comment, it borrows
all attributes from the last pattern with explicit phpdoc comment.

A good usage example of this feature is demonstrated in tests.

Made named rules enabled ("allowed") by default as well.
Whey can be disabled with -exclude-checks option if desired.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>