linter: add keywordCase checker
Unlike the previous implementation, this one does no
adhoc parsing and uses token positions info instead.
For statements and expressions where the token of
interest is the first token, everything is as simple
as checking string prefix. For tokens like "as" inside
foreach statement, we will need to do something more
intricate, since nothing is saved even inside freefloatings.
We also can't check for end* tokens with this approach.

But all in all, it implements ~90% of checks.

Updates #136

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>