src/linter: don't do Fatal if filename is a symlink
We used filepath.Walk before, now we use godirwalk.Walk.

filepath.Walk does not follow symlinks, but it does
accept it as a root argument without an error.
godirwalk.Walk can traverse symlinks with FollowSymbolicLinks=true,
but we don't use it. It will give an error if root is
a symlink, so we avoid calling Walk() on them.

This fix attempts to restore the old behavior.

Also added logging of the actual traversal error to
make this error debugging easier in the future.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>