solver: fix trait members accessLevel issues
Since traits can't extend or implement anything,
it's 100% guarantee that if we find a symbol inside
a trait it comes from that exact trait.

We can use that to achieve PHP-like semantics of
traits access level.

Suppose we're looking for a symbol S in class C that
uses trait T. Class C doesn't have it so we fall back
into traits. If trait T has S, return it with
implClass=C, so it looks like C itself had it.

That way, we can check accessLevel using the unmodified
canAccess() method.

Fixes #283
Fixes #209

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>