solver: make findX functions return result structs
Return both TraitName and ClassName in both
FindMethod and FindProperty functions.

Result.ImplName() returns the name that refers
to the implementation type name.

Result.ClassName is a type name that should be
used when resolving the access levels.

Result.TraitName is set to a trait name if Result.Info is
derived from the trait. Result.ClassName would be
a name of type that "uses" a trait.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>