add NodeSet type and start implementing dupCond check
NodeSet is a set-like data structure that only has 1 method - Add.

Add either adds an element and returns true or it returns false
if element is already in the set. The complexity is amortized to O(1)
due to the map[string] usage, but creating a map key is somewhat expensive.
There is an optimization for small N to make NodeSet alloc-free.

Moved FmtNode from linter package to astutil to avoid circular deps.

As an example of how we can use NodeSet there is a new dupCond check
that checks switch statement case clauses for duplicated entries.
It's planned to also check if-else chains, but it can wait until we
parse "elseif" and "else if" in the same way.

Refs #365

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>