Parse `} else if { ... }` as `} elseif { ... }`
ElseIf nodes that were created as a product of this merging
have Merged field set to true.

stmt.If.Else.Stmt now should never be stmt.If. All chained
elseifs are pushed to stmt.If.ElseIf slice.

To verify that it's working as intended, dupCond diagnostic is extended
to check for duplicated if/elseif conditions.

Fixes #365

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>