linter: validate @see references
phpdocRef is slightly more picky than PhpStorm reference
resolving as it does not permit referencing a type from
another namespace by incorrect relative name.

On the other side, if PhpStorm can't resolve the link,
phpdocRef will give a warning for sure.

The intention is to make it easier to keep @see code
references precise and up to date.

New warnings in flysystem are due to the relative namespace
reference is used: League\Flysystem\ReadInterface should be
\League\Flysystem\ReadInterface (with leading \).
PhpStorm can't resolve this reference either.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>