cmd/embeddedrules: embed rules into NoVerify binary
Added ternarySimplify checker as an example.

As before, named dynamic rules are disabled by default.
This can be changed for embedded rules in future.

This change also adds a little optimization:
if named rule is not enabled, don't load it.
This can save us from doing excessive work if
we're not going to report anything for the disabled
rule anyway.

To properly run tests for embedded rules, "go generate"
should be executed first for `./src/cmd` package.

	go generate ./src/cmd
	go test -v -run Golden/embeddedrules ./src/linttest

Fixes #461
Fixes #460

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>