linter: don't recurse when matching rules
phpgrep Find() method is recursive.
It tries to match at the top leven, then it traverses
the AST and calls the callback for every matching node.

A new tests (added in this change) resulted in 2 warnings
instead of 1 due to this.

It's not enough to do "return false" to stop the traversal.

More refactoring can be done later. This change goal
is to fix a bug.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>