phpgrep: add tracing tests + fix some old tests
Tracing tests help to understand and evaluate the
execution of the matcher for a given pattern+input pairs.

Tracing is guarded by the build tag so it should not
affect the performance in non-testing builds.

We ignored parsing errors, so some tests were skipped
due to that. Now parsing errors are checked.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>