meta: fix precision issue for mutable type maps (#481)
Append(n TypesMap) case for mutable maps had zero
checks for the precision. This made mutable maps
keep precision bit even if the imprecise map was appended.

We mark all mutable maps that are being appended
to as imprecise because we're lacking a good control
flow support (several open issues with how we deal with
if statements, etc.) and exprtype tests are not
exhaustive on the type precision yet.
More investigation is needed.

exprtype tests were fixed along the way.
If typesmap comes from a *scope.Var and it happens
to be mutable, it can change over the course of
the analysis. We need to do a Clone() to avoid that.

Note that TypesMap.Immutable() is not enough as it
only works if the map ownership was transferred.
In our case, it's not true as *scope.Var is still
the owner and this may result in our immutable
map being mutated after the exprTypeCollect is done.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>