phpdoc: fix phpdoc.Parse()
Since there can be multiple type annotations in a phpdoc
block, we need to do a Type.Clone() to avoid previously
parsed types corruption.

We could avoid that if we were processing phpdoc
directives one-by-one in a stream fashion, but since
we're eagerly converting all directives to a slice,
we can't avoid these allocations.

Fixes #497

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>