linttest: shuffle PHP input files
In our tests we don't do concurrent indexing/linting which
makes it harder to catch undesired processing order dependencies.

With randomized indexing/linting we're getting closer to the
real world cases where order of parsing is not guaranteed to
be the same between executions.

Test failure prints a seed that was used.
Using that value with $TEST_SEED environment variable
makes it possible to easily reproduce the failing state.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>