example,wasm: make go build ./... succeed
For wasm, we need all files to be excluded by the
build constraint (tags) to make go build/install/etc
pass on the repository root.

We also had leftovers of the main package in php/parser
which created a duplicated main package entry in the tree.

Another touch is fixing the example package.
"make check" will now run example/custom tests,
so hopefully we'll never break the example again.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>