linter: Added new checks for checking "dupArrayKeys" (close VCCOM#325)
Now duplicates are checked for the following:

1. Strings
2. Integers
3. Real numbers
4. Constants
5. Class constants
6. Variables
7. Access to array elements
8. Addition operations
9. Subtraction operations
10. Multiplication operations
11. Division operations
12. Concatenation

Duplicates in expressions are found only when the expressions are completely identical. However, expressions with two operands will be equal, even if the operands are reversed.

Real numbers in the keys are rounded to the nearest lower integer, so the keys 14.6 and 14.5 will be converted to 14 and a warning about duplicate keys will be issued.

See regression_test.go for example code with duplicates.

Signed-off-by: Group i582 <51853996+i582@users.noreply.github.com>