linter: Update checks for checking "dupArrayKeys" (close VCCOM#325)
Now the expressions do not have to be completely identical!, for example, if the expressions differ only in the order of the operands (and the operator is commutative), then a duplicate will be found.

Generally speaking, if two expressions give the same result, they will be found and marked as duplicates.

Signed-off-by: Group i582 <51853996+i582@users.noreply.github.com>