linter: Refactoring "dupArrayKeys" checks
Improved search algorithm, complexity changed from O(n^2) to O(n).

Fixed a bug in a situation where the keys contained expressions that were not processed into a hash function, and because of this, false positives occurred.

Fixed a bug in the processing of numbers in different number systems.

Added processing of 'PropertyFetch'.

Removed helper functions.

Added new tests.