cmd: replace the global flag vars with an explicit context
It's a small step, but it's better to do it now than never.

The first benefit that we get is that now it's obvious that
every InitEmbeddedRules call from the linttest package also
modified cmd package global state. It's not an error, but
it's not a pleasant side effect.

The next step would be to turn all linter.* vars from the
config.go file into a struct that would them and pass it
to the root walkers.

Updates #523

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>