cmd: fix git mode for the gitRepoComputeReportsFromCommits() case (#532)
Code was assigning to a wrong field so git.Zero was not replaced
by the "master" string. This led to a panic that is quite hard to
reproduce outside of the real-world environment.

To make this mistake less likely in future, I removed fields
from the struct and store these values as local values as they're
only needed inside that function scope.

gitRef looks unused. It was unused before the refactoring as well.
Maybe it was used before but we broke something a long time ago?
This calls for another investigation.

Refs #513

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>