solver: improve `self` handling in traits
When `self` is used inside a trait T, we expand that `self` to T
during the indexing phase (so metadata contains T instead of `self`).

This is not correct as T can't be a valid type on its own and `self`
always refers to a class C that embeds T.

Since T is not a valid type name, but it's known to be a trait name,
we can safely replace T with C during the types solving.

In contrast to the late static bindings resolution, we do that
replacement right away using the current scope className.

I also fixed a bug in WStaticPropertyFetch case. It was passing
class variable instead of className down the resolution chain
which led to invalid late static binding results.

Fixes #533

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>