implement baseline linting mode (#530)
-output-baseline is a new output mode that will produce baseline
file instead of normal reports. This option is used to generate
a suppression profile.

-baseline option enables baseline mode and uses a provided
suppression profile to filter out the output.

-conservative-baseline flag disables hashing of the prev and next
lines of the source code line being reported. This can be useful
if we want the behavior that is closer to the original diff mode,
but it may be more prone to false negatives. We need to collect
some stats on diagnostics and see which ones of them produce
more hash collisions on the real projects (can be solved by
making their report message more detailed and dynamic).

Baseline file is stored in JSON format and has several
size optimizations in 3 metrics:

	1. Line width
	2. Line numbers
	3. Overall file size

(1) and (2) are important to ensure that a file is openable
in most text editors. It also makes a file more git-friendly.

Warnings are converted to a 64-bit hash using the following fields:

	* Filename (only base)
	* Previous source line (in non-conservative mode only)
	* Start source line
	* Next soure line (in non-conservative mode only)
	* Warning message (includes a check name)
	* Warning scope (global/func/method name)

Updates #525

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>