meta: handle class and trait collisions
Like with functions, choose a "larger" definition.

This makes the linter output more stable as traversal
order will not influence indexing results.

We still may want to store slices of entries instead of only
one, but that's a bigger task and would require much more changes.

Output stability is required to make the baseline mode usable.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>