linter: add compliant/non-compliant code examples for every checker doc
linter.CheckerInfo gets 2 new fields: Before and After.
Before is a non-compliant code example (before the fix).
After is a compliant code example (after the fix).

A new package, lintdoc, provides utilities to render
check info in a human-readable format. It's meant to be
used in NoVerify as well is in custom NoVerify-based tools
that want to work with checks metadata.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>