cmd: add -gitignore flag (default=false)
When -gitignore is set, NoVerify tries to respect gitignore
files and exclude ignored files from the analysis.

During the startup, we try to find a git repository root
by going up until we find a folder with .git or encounter
the FS root. We save all gitignore files we find along the way.

gitignore can't be used outside of the valid git repository.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>