linttest: add e2e layer to golden tests
Some refactoring was required in order to make the work easier.

linter.Report structure now exports all its fields to
make it possible to write report formatting function
outside of the linter package.

The new flag -allow-all-checks is used in e2e tests to enable
all checkers without a need to enumerate all of them.

How end-to-end tests is implemented:

- If all golden tests pass, we run all of them again,
  but using the NoVerify binary that will be compiled
  as a part of the test.

- When -short flag is specified, e2e tests are skipped
  to reduce the time required to pass all tests.

- NoVerify binary is compiled with -race flag. It makes
  it run noticeably slower, but we get a chance to catch
  some race conditions during the testing.

Plans for the future:
- Test with cache as well.
- Add baseline mode tests.

Fixes #589

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>