linter: ignoring the maximum length for suggestions
If some rule sets a value for @fix, then a situation may arise that when replacing a string truncated to 60 characters will be placed, which will lead to syntax errors.
Therefore, for output to the console, if the proposed fix is longer than 60 characters, it is truncated, but in case of a fix it remains the original.

Fixes #598