linter: added checks for functions
Added check for bare returns for functions with non-void return types.
Added check for the absence of return statement in function with @return annotation, which is not equal to void.

For this, the storage of CurrentFunction has been improved.
CurrentFunction now stores the fully qualified name of the function with its class and namespace. A separate CurrentMethod field was created to store the names of the methods.