linter: start linting and indexing separation
As a first step, we move some simple checks to the
blockLinter type that runs only on linting phase
(so it doesn't have to guard against IsIndexingComplete).

There are some checks that are hard to be separated from
the walker, but it's a matter of time until we get to them.

As of now, new checks should be first attempted to be implemented
in blockLinter.

Additional changes to make block.go file smaller:

	* andWalker type goes to the separate file (and_walker.go)
	* some methods became functions and moved to block_utils.go

Changes in golden test are due to the different reports order
after the refactoring.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>