all: AST->IR transition
Since it's almost impossible to create a good IR from
the start as it would require a lot of time and efforts,
we make a conservative transition.

We copy all AST types into our IR package and
create a mapping function irgen.Convert(AST)->IR.

When we want to modify IR, we change our own type
definitions and update the irgen pacakge.

When we want to update the php-parser, usually,
only irgen package would require to be updated.

Next steps:
* Use upstream php-parser lib as irgen input
* Remove our php-parser modified fork
* Improve IR-related code (as it was mostly generated)
* Stop depending on the freeloating/positions as well?
* Move php/astutil -> ir/irutil

The main difference right now is that all IR nodes
live in ir package. This is usually easier to maintain,
especially if we want to rely on code generation later.

All packages rely on IR now (there are few exceptions,
like position and freefloating packages, but it's not
relevant right now).

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>