ir: rework Name nodes
Merge Name+FullyQualifiedName and store all parts as a
combined string.

When we need to access name individual parts, we can
do so via Name methods. They are zero-alloc and are
somewhat cheap to compute (not more expensive than
having to turn old name parts to string for sure and
compare them via NameEquals instead of just ==).

FQN are expressed with Name.Value with leading `\`.
Normal names do not have leading `\`.

RelativeName is converted into Name with namespace expansion
during the irgen phase, so we eliminate that node type as well.
Resolved RelativeNames are FQNs.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>