ir: simplify argument lists (and remove ArgumentList node)
This is done to simplify arguments handling.
We can't store arguments as []*Argument since it will
require us to duplicate code that accepts []Node
(and we don't have generics yet).

As a workaround, all nodes with arguments get Arg(i) method
that returns ith argument type-asserted to *Argument.

So we can do:
	call.Arg(i)

Instead of:
	call.ArgumentList.Arguments[i].(*ir.Argument)

This does look like an improvement.

To avoid information loss, freefloating from the ArgumentList
is stored inside the call-like node itself. So if we ever
need it, it's there.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>