linter: improve dup array key messages
Print both duplicated key syntax (as in the source file) and value,
but do not print the value if it's identical to the syntax string.

Since ir.String can contain non-printable chars like \n, we print
values with %q, so \n will be displayed as "\n" and will not end
up creating an actual newline in the output.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>