rules: add multi-pattern rules support via "any" blocks
For the detailed overview, see #709.

This patch introduces a new dynamic rules concept: special
rule grouping blocks.

	// Matches if any of the provided patterns matches.
	any: {
	  $a;
	  $b;
	  // ...
	}

	// Matches if we can match $a followed by $c, ...
	seq: {
	  $a;
	  $b;
	  // ...
	}

To avoid duplicated label warnings in IDE, we permit any_ and seq_ as
prefixes, so any_eq3 is interpreted in the same way as any:

	/** @warning === has higher precedence than | */
	any_eq3_bitor: {
	  $_ === $_ | $_;
	  $_ | $_ === $_;
	}

The implementation is rough on some places due to the experimental
state of this feature. The prototype (proto) based algorithm makes
it possible to pass function-defined properties later on if we
feel like more metadata reuse is required.

Fixes #709
Fixes #276

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>