linter: add a Worker concept
A worker is a single goroutine that reads from a filename
channel and runs analysis over them.

We can keep some shared state between the RootWalker's that
are executed on the same goroutine if we'll express
that worker explicitly as an object.

The new Worker type expresses that linter handle that
stores the state that can be reused.

Right now it only contains the phpdoc types parser that
has an internal pool of expressions. If it's used several
times, it becomes zero-alloc. If we reuse it for all RootWalker's
inside for a worker, we'll cut a few redundant allocations.

When php-parser will be updated an *Parser object will become
reusable, we can store PHP parser inside the worker state as well.

I also have some plans to make irconv package run faster.
In order to do that, keeping the converted can be beneficial.

Old linter.ParseContents and linter.IndexFile API remains,
but it's deprecated in favor of Worker-based API that is more efficient.

Right now we use Worker-based API in normal linter mode as well
as in linter tests (linttest pkg). Language server continues
to use the old API.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>