ir/irconv: do AST->IR via converter API
Introduce the Converter object that will be used everywhere
where we want to do AST->IR.

The benefit is that we can clearly see where that converter
can be reused and where it can be used as a one-off object.

The ConvertRoot and ConvertNode functions are removed
to avoid API duplication.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>