linter: move bitwiseOps to dynamic rules
As a bonus, & and | that are used instead of && and || now
can be auto-fixed.

We also start to check for &, | and ^ applied to non-numeric
types. In our test corpus it only triggers in math library
that does that on purpose (bitwise work on strings).
Since sometimes this is a desired behavior, these extra
patterns have MAYBE severity level.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>