tests: make e2e tests ~x6 faster
Instead of doing a full stubs loading, use our dependencies
list from the golden tests to hint the linter which embedded
stubs need to be loaded.

With this, we significantly reduce the amount of work linter
needs to do before running the inspections.

On my machine, golden tests run for 40-50 seconds.
After this patch I get 5-8 seconds.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>