ir: eagerly parse phpdoc in irconv
The irconv package will now attempt to parse phpdoc for common
nodes like function/method/class statements if
Converter was created with phpdoc types parser.
(This API may change in the future!)

We need to do eager parsing in 99% of cases.
The only case where we don't want to do it
are phpgrep-like utils as well as some php-guru
parts that are not interested in phpdoc comments at all.

All packages that were parsing raw strings are changed to
use []phpdoc.CommentPart directly.

I returned irconv.ConvertNode() helper function to
make it easier to make AST->IR conversion in places
where we are too lazy to reuse the converter or
we don't care about their performance (some of these
places can be fixed later if we ever want to).

The main benefit of the eager phpdoc parsing is that
custom (external) checkers don't have to repeat all
phpdoc parsing efforts. Right now these checkers need
to re-parse class/method/function phpdoc strings if
they want to do something fancy with them. With this
change, we can access the parsed structures directly
from both internal and external checkers.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>