checkers: fix regexpsimplify handling of named capture alt forms
Named capture always printed like ?P<name>, but it can have 2
more forms: ?<name> and ?'name'.

The quoted format is extra tricky as we always suggest '-quoted
replacements; we give up if we find this form to avoid invalid
suggestions. It's possible to suggest "-quoted literals in
the future, but it may require more changes (and tests).

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>