constfold: refactoring related to constants
Eval function interface changed to (meta.ConstValue, bool).
Motivation:
At the moment, checking for the fact that the folding of
constants was not successful is done as follows:

constFolded := constfold.Eval(norm.st, e)
if constFolded.Type == meta.Undefined {
   // some code
}

Which is not very pretty.
The advantage of this interface was that it is quite
convenient to write code for different expressions in one
line, for example:

return Neg(Eval(st, e.Expr))

As a result, all the logic was moved to the eval function
with the old interface, and the Eval function was created
for the new interface, which checks Type == meta.Undefined
in itself and returns the value and the required flag.

The check now looks like this:

constFolded, ok := constfold.Eval(norm.st, e)
if !ok {
   // some code
}

In the eval function, some direct constructs of the
meta.ConstValue structure has been replaced with
meta.New[Type]Const functions.

meta.New[Type]Const function names have been normalized.

A comment has been added for the ConstValue.Get[Type],
ConstValue.New[Type]Const, ConstValue.IsEqual functions
and ConstValue structure.

The ConstInfo structure has been moved to a better
location – metainfo.go.