workspace: fixed bug when untracked binary was added in git mode and this led to panic
In git mode, in the case of untracked files, the gitParseUntracked function is called, which passes the list of files to the workspace.ReadFilenames function, in which the readFilenames function is called for each file.

The problem is that if not a directory is passed, then the following code is executed

if !st.IsDir() {
  if filter.IgnoreFile(filename) {
    return
  }

  ch <- FileInfo{Filename: filename}
  return
}

Where there is no extension check, which is present when traversing a folder below the function.
Let's add a check here and the error will go away.